Made UATs for userstory8
